+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/aio/_lro_async.py
@@ -7,6 +7,7 @@ import base64
 import functools
 import json
 from typing import Optional, Any
+from msrest import Deserializer
 from azure.core.exceptions import HttpResponseError
 from azure.core.polling import AsyncLROPoller
 from azure.core.polling.base_polling import OperationFailed, BadStatus
@@ -120,7 +121,9 @@ class AsyncAnalyzeHealthcareEntitiesLROPollingMethod(
     def last_modified_on(self):
         if not self._current_body:
             return None
-        return self._current_body.last_update_date_time
+        return self._current_body.last_updated_date_time or Deserializer().deserialize_iso(
+                self._current_body.additional_properties["lastUpdateDateTime"]
+            )  # backwards compatibility
 
     @property
     def id(self):
@@ -307,7 +310,9 @@ class AsyncAnalyzeActionsLROPollingMethod(TextAnalyticsAsyncLROPollingMethod):
     def last_modified_on(self):
         if not self._current_body:
             return None
-        return self._current_body.last_update_date_time
+        return self._current_body.last_updated_date_time or Deserializer().deserialize_iso(
+                self._current_body.additional_properties["lastUpdateDateTime"]
+            )  # backwards compatibility
 
     @property
     def total_actions_count(self):