+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze.py
@@ -1814,7 +1814,6 @@ class TestAnalyze(TextAnalyticsTest):
             actions=[
                 AnalyzeHealthcareEntitiesAction(
                     model_version="latest",
-                    fhir_version="4.0.1"
                 )
             ],
             show_stats=True,
@@ -1828,5 +1827,4 @@ class TestAnalyze(TextAnalyticsTest):
                     assert res.error.code == "InvalidDocument"
                 else:
                     assert res.entities
-                    assert res.fhir_bundle
                     assert res.statistics