+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_async.py
@@ -1928,7 +1928,6 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                 actions=[
                     AnalyzeHealthcareEntitiesAction(
                         model_version="latest",
-                        fhir_version="4.0.1"
                     )
                 ],
                 show_stats=True,
@@ -1945,5 +1944,4 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                         assert res.error.code == "InvalidDocument"
                     else:
                         assert res.entities
-                        assert res.fhir_bundle
                         assert res.statistics